+++ ChangeLog.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 
 - improvements to REPL for printing to terminal
+- function `io/list-files-glob` to list files with a glob pattern